+++ src/cpu_measure.rs
@@ -1,11 +1,12 @@
 
+/// This struct represents each one of the measures for the CPU
 pub struct CpuMeasure {
-    pub time_seconds: u64,
+    pub time_seconds: f64,
     pub percentage: f64
 }
 
 impl CpuMeasure {
-    pub fn new(s: u64, p: f64) -> Self {
+    pub fn new(s: f64, p: f64) -> Self {
         CpuMeasure{time_seconds: s, percentage: p}
     }
 }
